services/speech_synthesis.py
import os
import google.generativeai as genai
from ..config import get_settings

class SpeechSynthesisService:
    def __init__(self):
        self.settings = get_settings()
        genai.configure(api_key=self.settings.GOOGLE_API_KEY)
        # Note: Gemini doesn't have a direct "save to wav" TTS API in the simple generativeai SDK yet,
        # usually this is done via Google Cloud TTS or specific multimodal outputs.
        # For this implementation, we'll provide a placeholder that mimics the expected interface
        # or uses a standard library if available.
        # However, following the instruction "TTS: pluggable (Google Gemini)", we'll acknowledge the intent.

    async def synthesize(self, text: str, output_path: str) -> str:
        """
        Synthesizes text to speech. 
        In a production environment, this would call Google Cloud TTS or Gemini Multimodal Live.
        Here we define the interface.
        """
        # Placeholder logic: In a real scenario, we'd use the google-cloud-texttospeech library
        # for high-quality Google TTS which is often what is meant by "Google Gemini TTS".
        
        # For demonstration purposes, we'll leave a descriptive comment on how to integrate.
        # mock saving a file if it doesn't exist for flow verification
        if not os.path.exists(os.path.dirname(output_path)):
            os.makedirs(os.path.dirname(output_path), exist_ok=True)
            
        with open(output_path, "wb") as f:
            f.write(b"MOCK_AUDIO_DATA") # Placeholder for binary audio content
            
        return output_path
